% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactFunc-out.R
\name{asNormalFunc}
\alias{asNormalFunc}
\title{Convert a cacheable function to a normal function}
\usage{
asNormalFunc(f)
}
\arguments{
\item{f}{A function built with \code{\link{reactFunc}}.}
}
\value{
Return a function having similar behavior with the input function.
}
\description{
Functions built with \code{\link{reactFunc}} can be converted to a normal
version so that you do not need to replicate yourself.
}
\examples{
refunc <- reactFunc(ARGV = alist(x = , y = ),
    ans = a() * b(),
    a = x + 1,
    b = y - 3,
    ret = ans()
)
normfunc <- asNormalFunc(refunc)
normfunc
identical(refunc(6,9), normfunc(6,9))
}
